+++ presets/index.js
@@ -1,8 +1,8 @@
 const antd_vue_v1 = require('./ant-design-vue-v1');
 const antd_vue_v3 = require('./ant-design-vue-v3');
 const antd_v4 = require('./antd-v4');
-const element_plus = require('./element-plus');
-const element_ui = require('./element-ui');
+const element_plus = require('./element-plus-v2');
+const element_ui = require('./element-ui-v2');
 const pinia_v2 = require('./pinia-v2');
 const react_dom_v17 = require('./react-dom-v17');
 const react_dom_v18 = require('./react-dom-v18');
@@ -13,7 +13,7 @@ const react_router_v6 = require('./react-router-v6');
 const react_v17 = require('./react-v17');
 const react_v18 = require('./react-v18');
 const redux_v5 = require('./redux-v5');
-const vue_composition_api = require('./vue-composition-api');
+const vue_composition_api = require('./vue-composition-api-v1');
 const vue_router_v4 = require('./vue-router-v4');
 const vue_v2 = require('./vue-v2');
 const vue_v3 = require('./vue-v3');
